# openapi.yaml
openapi: 3.0.3
info:
  title: tr-engine API
  version: 0.9.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Backend service for trunk-recorder data ingestion and querying.
    Provides REST APIs for accessing radio system data, calls, talkgroups,
    units, and recorder state.

    ## Authentication

    All endpoints (except `/health` and `/auth-init`) require a bearer token
    via the `Authorization: Bearer <token>` header or `?token=` query param.

    **Two-tier auth (optional):** When `WRITE_TOKEN` is configured, mutating
    operations (POST, PATCH, PUT, DELETE) require the write token. Read
    operations (GET) accept either the read token (`AUTH_TOKEN`) or write
    token. If `WRITE_TOKEN` is not set, all operations accept `AUTH_TOKEN`
    (backwards compatible).

    The `/auth-init` endpoint returns only the read token — the write token
    is never exposed by any endpoint.

    ## Data Model: Systems and Sites

    The data model separates **logical radio networks** from **recording
    sites**:

    - **System** (`system_id`): a logical radio network. For P25, this
      is the WACN-level network identified by `(sysid, wacn)`. For
      conventional, it is identified by `(instance_id, sys_name)`.
      Talkgroups and units belong to the system level because they are
      shared across all sites on the same P25 network.

    - **Site** (`site_id`): a recording point within a system. Each
...
      specific site.

...
    ```
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /unit-events:
    get:
      operationId: listUnitEventsGlobal
      summary: List unit events (system-wide)
      description: |
        Returns unit events across an entire system with JOINs for display
        names. Supports multi-type filtering and optional emergency filter.
        Default lookback is 1 hour; max time range is 24 hours.

        **At least one of `system_id` or `sysid` is required.** Returns 400
        if neither is provided.
      tags: [units]
      parameters:
        - name: system_id
          in: query
          description: |
            Filter by internal system ID (comma-separated for multiple).
            Alias "systems" also accepted. **Required** unless `sysid` is provided.
          schema:
            type: string
            example: "1,2"
        - name: sysid
          in: query
          description: |
            Filter by P25 sysid hex (comma-separated for multiple).
            Alias "sysids" also accepted. **Required** unless `system_id` is provided.
          schema:
            type: string
        - name: unit_id
          in: query
          description: Filter by unit radio ID (comma-separated for multiple). Aliases "units", "unit_ids" also accepted.
          schema:
            type: string
            example: "924003,104"
        - name: type
          in: query
          description: |
            Comma-separated event types to include (e.g. `join,on,off`).
            If omitted, all event types are returned.
          schema:
            type: string
        - name: tgid
          in: query
          description: Filter by talkgroup ID (comma-separated for multiple). Alias "tgids" also accepted.
          schema:
            type: string
            example: "9178,5344"
        - name: emergency
          in: query
          description: Filter by emergency flag
          schema:
            type: boolean
        - $ref: "#/components/parameters/startTime"
        - $ref: "#/components/parameters/endTime"
        - name: sort
          in: query
          description: |
            Sort field with optional `-` prefix for descending.
            Allowed: `time`, `unit_rid`, `tgid`, `event_type`. Default: `-time`.
          schema:
            type: string
            default: "-time"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitEventListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ----------------------------------------------------------
  # Unit Affiliations (live in-memory state)
  # ----------------------------------------------------------
...
  /calls/{id}/transcription:
    get:
      operationId: getCallTranscription
      summary: Get primary transcription for a call
      description: Returns the current primary transcription with unit-attributed word timestamps and segments.
      tags: [transcriptions]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transcription"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: submitTranscriptionCorrection
      summary: Submit transcription
      description: |
        Creates a new primary transcription variant for a call. Defaults to
        source "human" for backward compatibility, but callers can supply
        source, provider, language, and pre-built word/segment data for
        source-side or programmatic transcriptions.
      tags: [transcriptions]
      parameters:
        - $ref: "#/components/parameters/callId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  description: Transcription text
                source:
                  type: string
                  description: Transcription source. Defaults to "human" if omitted.
                  example: human
                provider:
                  type: string
                  description: Provider identifier (e.g., "client-whisper", "elevenlabs")
                  example: client-whisper
                language:
                  type: string
                  description: Language code
                  example: en
                words:
                  type: object
                  nullable: true
                  description: |
                    Optional pre-built word/segment data. Same structure as the
                    `words` field in the Transcription schema.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  call_id:
                    type: integer
                  source:
                    type: string
                    example: human
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /calls/{id}/transcriptions:
    get:
      operationId: listCallTranscriptions
      summary: List all transcription variants for a call
      description: |
        Returns all transcription variants (auto, human, LLM) for a call.
        The primary variant is marked with `is_primary: true`.
      tags: [transcriptions]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  transcriptions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transcription"
                  total:
                    type: integer
        "404":
          $ref: "#/components/responses/NotFound"
  /calls/{id}/transcribe:
    post:
      operationId: transcribeCall
      summary: Queue call for transcription
      description: Queues a call for transcription (or re-transcription). Returns 202 if accepted.
      tags: [transcriptions]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "202":
          description: Accepted — queued for transcription
          content:
            application/json:
              schema:
                type: object
                properties:
                  call_id:
                    type: integer
                  status:
                    type: string
                    enum: [queued]
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: Transcription queue full or not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /calls/{id}/transcription/verify:
    post:
      operationId: verifyTranscription
      summary: Verify a transcription
      description: Marks the call's current transcription as human-verified.
      tags: [transcriptions]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusChangeResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /calls/{id}/transcription/reject:
    post:
      operationId: rejectTranscription
      summary: Reject a transcription
      description: Resets the call's transcription_status to "reviewed" (unverified).
      tags: [transcriptions]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusChangeResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /calls/{id}/transcription/exclude:
    post:
      operationId: excludeFromDataset
      summary: Exclude call from training dataset
      description: Marks a call as permanently excluded from the training dataset.
      tags: [transcriptions]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusChangeResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /transcriptions/search:
    get:
      operationId: searchTranscriptions
      summary: Full-text search across transcriptions
      description: |
        Searches transcription text using PostgreSQL full-text search.
        Results include call context (talkgroup, system, timing) so no
        follow-up lookups are needed.
      tags: [transcriptions]
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
        - name: system_id
          in: query
          description: Filter by system ID(s), comma-separated. Alias "systems" also accepted.
          schema:
            type: string
        - name: tgid
          in: query
          description: Filter by talkgroup ID(s), comma-separated. Alias "tgids" also accepted.
          schema:
            type: string
        - name: site_id
          in: query
          description: Filter by site ID(s), comma-separated. Alias "sites" also accepted.
          schema:
            type: string
        - name: start_time
          in: query
          description: Start of time range (RFC 3339)
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          description: End of time range (RFC 3339)
          schema:
            type: string
            format: date-time
        - name: primary_only
          in: query
          description: When true, only return the primary transcription for each call
          schema:
            type: boolean
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionSearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /transcriptions/queue:
    get:
      operationId: getTranscriptionQueueStatus
      summary: Get transcription queue status
      description: Returns statistics about the transcription processing queue.
      tags: [transcriptions]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionQueueStats"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ----------------------------------------------------------
  # Call Upload
  # ----------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Pass token in the Authorization header:
        `Authorization: Bearer <token>`

  # ----------------------------------------------------------
  # Reusable Parameters
  # ----------------------------------------------------------
  parameters:
    callId:
      name: id
      in: path
      required: true
      description: Call record database ID
      schema:
        type: integer
        example: 48531
...
    startTime:
      name: start_time
      in: query
      description: Start of time range (RFC 3339)
      schema:
        type: string
        format: date-time
    endTime:
      name: end_time
      in: query
      description: End of time range (RFC 3339)
      schema:
        type: string
        format: date-time
    limit:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 1000
    offset:
      name: offset
      in: query
      description: Page offset (number of results to skip)
      schema:
        type: integer
        default: 0
        minimum: 0
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Forbidden — valid read token but write token required for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    EventType:
      type: string
      enum: [on, off, join, call, end, data, ans_req, location, ackresp]
      description: |
        Unit event types from trunk-recorder MQTT:
        - **on**: unit registration (radio powered on / entering system)
        - **off**: unit deregistration (radio powered off / leaving system)
        - **join**: talkgroup affiliation (unit joins a talkgroup)
        - **call**: channel grant (voice call participation)
        - **end**: transmission end (per-unit end-of-transmission report)
        - **data**: data grant
        - **ans_req**: answer request
        - **location**: unit location update
        - **ackresp**: acknowledge response
...
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: Machine-readable error code for programmatic handling
          example: not_found
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
            - invalid_body
            - invalid_parameter
            - invalid_time_range
            - query_failed
            - ambiguous_id
            - duplicate
            - request_timeout
        error:
          type: string
          example: Resource not found
        detail:
          type: string
          description: Additional context when available
...
    UnitEvent:
      type: object
      description: A unit event (registration, affiliation, call activity, etc.)
      required: [id, event_type, unit_rid, unit_id, time]
      properties:
        id:
          type: integer
          example: 1
        event_type:
          $ref: "#/components/schemas/EventType"
        time:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"

        # System context
        system_id:
          type: integer
          description: Database system ID
          example: 1
        system_name:
          type: string
          description: System display name
          example: "Butler/Warren P25"

        # Unit context (embedded to avoid lookup)
        unit_rid:
          type: integer
          description: "P25 Radio ID (legacy name; prefer unit_id for consistency)"
          example: 1234567
        unit_id:
          type: integer
          description: "Radio unit ID (alias for unit_rid, consistent with other endpoints)"
          example: 1234567
        unit_alpha_tag:
          type: string
          description: Unit display name (embedded to avoid lookup)
          example: Engine 1

        # Talkgroup context (embedded to avoid lookup)
        tgid:
          type: integer
          example: 1001
        tg_alpha_tag:
          type: string
          description: Talkgroup short-form display name
          example: "09-8L Main"
        tg_description:
          type: string
          description: Human-readable talkgroup description
          example: Middletown Police Dispatch

        # Internal references
        instance_id:
          type: string
          description: Trunk-recorder instance that reported this event
          example: trunk-recorder
        metadata_json:
          type: object
          additionalProperties: true
...
    TranscriptionSource:
      type: string
      enum: [auto, human, llm]
      description: |
        - **auto**: automated transcription (e.g., Whisper)
        - **human**: human correction/transcription
        - **llm**: LLM-generated correction
    Transcription:
      type: object
      description: A transcription of a call's audio content with unit attribution
      required: [id, call_id, text, source]
      properties:
        id:
          type: integer
          example: 1
        call_id:
          type: integer
          example: 48531
        text:
          type: string
          example: Engine 5 responding to the scene
        source:
          $ref: "#/components/schemas/TranscriptionSource"
        is_primary:
          type: boolean
          example: true
        confidence:
          type: number
          nullable: true
          minimum: 0
          maximum: 1
          example: 0.95
        language:
          type: string
          example: en
        model:
          type: string
          example: deepdml/faster-whisper-large-v3-turbo-ct2
        provider:
          type: string
          example: whisper
        word_count:
          type: integer
          example: 6
        duration_ms:
          type: integer
          description: Transcription processing time in milliseconds
          example: 1500
        created_at:
          type: string
          format: date-time
        words:
          type: object
          description: |
            Unit-attributed word timestamps and segments. The `words` array
            contains every word with timing and the radio unit that said it.
            The `segments` array groups consecutive words by the same unit.
          properties:
            words:
              type: array
              items:
                $ref: "#/components/schemas/AttributedWord"
            segments:
              type: array
              items:
                $ref: "#/components/schemas/TranscriptionSegment"
    AttributedWord:
      type: object
      description: A word with timing and unit attribution
      required: [word, start, end, src]
      properties:
        word:
          type: string
          example: Engine
        start:
          type: number
          description: Start time in seconds from audio start
          example: 0.12
        end:
          type: number
          description: End time in seconds from audio start
          example: 0.45
        src:
          type: integer
          description: Unit/radio ID (0 if unattributed)
          example: 9130
        src_tag:
          type: string
          description: Unit alpha tag
          example: "Medic 83"
    TranscriptionSegment:
      type: object
      description: Consecutive words from the same radio unit
      required: [src, start, end, text]
      properties:
        src:
          type: integer
          description: Unit/radio ID
          example: 9130
        src_tag:
          type: string
          example: "Medic 83"
        start:
          type: number
          example: 0.12
        end:
          type: number
          example: 1.8
        text:
          type: string
          example: "Medic 83 transporting to Fort."
    TranscriptionSearchHit:
      type: object
      description: A search result with relevance score and call context
      properties:
        id:
          type: integer
        call_id:
          type: integer
        text:
          type: string
        source:
          type: string
        is_primary:
          type: boolean
        word_count:
          type: integer
        duration_ms:
          type: integer
        created_at:
          type: string
          format: date-time
        rank:
          type: number
          description: Search relevance score (higher is better)
        system_id:
          type: integer
        system_name:
          type: string
        tgid:
          type: integer
        tg_alpha_tag:
          type: string
        call_start_time:
          type: string
          format: date-time
        call_duration:
          type: number
          nullable: true
    TranscriptionSearchResponse:
      type: object
      required: [results, total, limit, offset]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/TranscriptionSearchHit"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    TranscriptionQueueStats:
      type: object
      properties:
        status:
          type: string
          enum: [ok, not_configured]
        pending:
          type: integer
          example: 15
        completed:
          type: integer
          example: 5000
        failed:
          type: integer
          example: 12
    StatusChangeResponse:
      type: object
      properties:
        call_id:
          type: integer
          example: 48531
        status:
          type: string
          example: verified
...
    UnitEventListResponse:
      type: object
      required: [events, total, limit, offset]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/UnitEvent"
        total:
          type: integer
          example: 100
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
...